Actually pay attention to -Wno-missing-include-dirs from gfortran.
Also include -Wno-missing-include-dirs with -fdec.

Note that this fix is defunct for gcc-trunk and should be changed.
There is already an upstream commit that attempts to fix this warning, but
still doesn't do the trick when -I appears before -Wno-missing-include-dirs
(or even sometimes when it does appear in the right order).